Added animation class to engine. Added suppot for alpha channel to engine. Added click animation when switch clicked, for visual feedback. Replaced text and button with original translucent versions. Fixed comparison between signed and unsigned int in keyboard.

+++ src/Keyboard.cpp
@@ -63,7 +63,7 @@ void Keyboard::Step()
 	if ( timer > 0 )
 		--timer;
 
-	if ( text.length() > maxStringLength and timer == 0 )
+	if ( (int)text.length() > maxStringLength and timer == 0 )
 		text.erase( text.end() - 1 );
 
 	if (curMouseDown and !isMouseDown)
